+++ tests/unittests/test_fitting.py
@@ -297,7 +297,7 @@ def test_custom_fitness(yaml, simulated, target, weighting):
                     100.0,
                 ]
             ),
-            88663.74498686475,
+            88431.18016117146,
         )
     ],
 )